+++ src/Pages/Accounts/Components/AccountSpotlight.js
@@ -1,4 +1,4 @@
-import React, { Fragment, useEffect } from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
 import styled from 'styled-components';
 import { useParams } from 'react-router-dom';
 import { useAccounts, useMediaQuery } from 'redux/hooks';
@@ -19,6 +19,8 @@ const AccountSpotlight = () => {
   const { accountInfo, accountInfoLoading, getAccountAssets, getAccountInfo } = useAccounts();
   const { addressId } = useParams();
   const { matches } = useMediaQuery(breakpoints.down('sm'));
+  const { matches: matchesLong } = useMediaQuery(breakpoints.down('xl'));
+  const [showPopup, setShowPopup] = useState(false);
 
   useEffect(() => {
     getAccountAssets(addressId);
@@ -30,10 +32,26 @@ const AccountSpotlight = () => {
     accountNumber = '--',
     accountType = '--',
     publicKeys = {},
+    publicKey = publicKeys.pubKey,
+    publicKeyType = publicKeys.type,
     sequence = '--',
     tokens = {},
   } = accountInfo;
-  const publicKey = publicKeys.signers ? publicKeys.signers[0] : '';
+
+  const popupNoteKeyType = {
+    visibility: { visible: showPopup, setVisible: setShowPopup },
+    icon: { name: 'HELP_OUTLINE', size: '1.7rem' },
+    method: ['click', 'hover'],
+    fontColor: 'FONT_WHITE',
+    data: [
+      {
+        title: 'Key Type:',
+        value: publicKeyType,
+      },
+    ],
+    titleMinWidth: '70px',
+    noteMinWidth: '70px',
+  };
 
   const summaryData = [
     { title: 'Address', value: matches ? maxLength(addressId, 20, 3) : addressId, copy: addressId },
@@ -42,8 +60,9 @@ const AccountSpotlight = () => {
     { title: 'Account Number', value: accountNumber },
     {
       title: 'Public Key',
-      value: matches ? maxLength(publicKey, 20, 3) : publicKey,
+      value: matchesLong ? maxLength(publicKey, 20, 3) : publicKey,
       copy: publicKey,
+      popupNote: popupNoteKeyType,
     },
     { title: 'Sequence', value: sequence },
     { title: 'Fungible Tokens', value: tokens.fungibleCount },